[CCUBE-2183][GZ] generated argtypes for animations
