+++ src/animations/types.ts
@@ -3,20 +3,28 @@ import type {
     ComponentTokenCSSVariableString,
 } from "../theme";
 
-export interface BaseAnimationProps {
+// @storybookSkipProps
+type BaseAnimationProps = {
     id?: string;
     "data-testid"?: string | undefined;
     className?: string;
-}
+};
 
-export interface CustomisedProps {
+/**
+ * Props for the animations component.
+ */
+export type AnimationProps = BaseAnimationProps & {};
+
+/**
+ * Props for any customisable animations component.
+ */
+export type CustomisableAnimationProps = BaseAnimationProps & {
+    /**
+     * The color that is to be applied on the animated elements
+     */
     color?:
         | string
         | ColourCSSVariableString
         | ComponentTokenCSSVariableString
         | undefined;
-}
-
-export interface CustomisableAnimationProps
-    extends BaseAnimationProps,
-        CustomisedProps {}
+};
