--- stories/animation/2-animations-customisable.mdx
@@ -1,33 +1,0 @@
-import { Canvas, Meta, Story } from "@storybook/addon-docs/blocks";
-import { DocNote } from "stories/storybook-common";
-import * as AnimationsCustomisableStories from "./2-animations-customisable.stories";
-import { CustomisedPropsTable } from "./props-table";
-
-<Meta of={AnimationsCustomisableStories} />
-
-# Customisable Animations
-
-<DocNote>
-    Customisation is only supported by the animations on this page.
-</DocNote>
-
-## LoadingDotsSpinner
-
-```tsx
-import { LoadingDotsSpinner } from "@lifesg/react-design-system/animations";
-```
-
-For this component, if no color is specified, it will use the
-`Animation["loading-dots-spinner-colour"]` token.
-
-<Canvas of={AnimationsCustomisableStories.CustomLoadingDotsSpinner} />
-
-## Component tokens
-
-This component can be customised via the `Animation` tokens.
-
-<Story of={AnimationsCustomisableStories.TokenCustomisation} />
-
-## Component API
-
-<CustomisedPropsTable />
