+++ stories/animations/props-table.tsx
